Fixed issue with flushing staging buffers as the stream shutsdown.
